Frameworks: hw key rebinding. Allow virtual keys to preload recents
Apps like LMT or similar which send hw key events as virtual key
cannot load new recents right now due of missing proper preload.
Well it was before as well broken the old recents was just not that
strict then the new one....but it should be due that this was one of the
main reasons why the old one was buggy. Bad preloading

whatever fix it and simply allow virtual keys which fire app switchter
key to fire a recents preload.

Change-Id: I09b96b50894af8a91c261e17c736a66ca377952d